small refactor to app and express
