+++ app/api/users/controller.js
@@ -1,6 +1,6 @@
 import Users from './model';
 
-import {handleError, respondWithResult} from '../../services/nodeResponse/apiResponses';
+import {handleError, respondWithResult} from '../../services/node/nodeResponse/apiResponses';
