+++ app/app.js
@@ -1,30 +1,7 @@
 import express from 'express';
-import webpack from './services/middlewares/webpack';
-import bodyParser from './services/middlewares/bodyParser';
-import views from './services/middlewares/views';
-import errors from './services/middlewares/errors';
-import logger from './services/middlewares/logger';
-import validator from './services/middlewares/validator';
-import router from './services/middlewares/router';
-import serveStatics from './services/middlewares/serveStatics';
-import api from './services/middlewares/api';
-import session from './services/middlewares/session';
-import passport from './services/middlewares/passport';
-
+import loadExpressMiddlewares from './services/express/middlewares/load';
 const app = express();
 
-logger(app);
-app.set('x-powered-by', false);
-app.set('view cache', true);
-webpack(app);
-serveStatics(app);
-bodyParser(app);
-session(app);
-views(app);
-validator(app);
-passport(app);
-api(app);
-router(app);
-errors(app);
+loadExpressMiddlewares(app);
 
 export default app;
