+++ app/client.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'flexboxgrid'; // load flexbox for grid system
 import './styles/custom-styles.scss';
-import './services/socket/client'; // connect to socket client
+import './services/node/socket/client'; // connect to socket client
 import AppWrapper from './Wrappers/App';
 import {render} from 'react-dom';
 const ROOT = document.getElementById('root');
