+++ app/services/express/middlewares/passport/index.js
@@ -1,4 +1,4 @@
-import auth from './../../../config/auth';
+import auth from '../../../../config/auth';
 import passport from 'passport';
 export default (app) => {
     app.use(passport.initialize());
