+++ app/services/express/middlewares/reduxState/index.js
@@ -1,8 +1,8 @@
-import configureStore from '../../../redux/store/store';
+import configureStore from '../../../../redux/store/store';
 // config redux store and save in locals - template will use it to render initialized app with server side state
 
-import {fountUser} from '../../../redux/actions/user';
-import {fetchPayments} from '../../../redux/actions/payments';
+import {fountUser} from '../../../../redux/actions/user';
+import {fetchPayments} from '../../../../redux/actions/payments';
 
 export default (req, res, next) => {
     let store = configureStore();
