+++ app/services/express/middlewares/setters/index.js
@@ -1,0 +1,4 @@
+export default (app) => {
+    app.set('x-powered-by', false);
+    app.set('view cache', true);
+}
