+++ app/services/express/middlewares/webpack/index.js
@@ -1,8 +1,8 @@
 import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
-import webpackConfig from '../../../../webpack.config';
-import {env} from '../../../config/env';
+import webpackConfig from '../../../../../webpack.config';
+import {env} from '../../../../config/env';
 import {initial} from 'lodash';
 
 if (env === 'test') {
