+++ app/services/node/socket/client.js
@@ -1,4 +1,4 @@
-import {host} from '../../config/env'
+import {host} from '../../../config/env'
 import io from 'socket.io-client';
 
 let socket = io.connect(host, {reconnect: true});
